tmdb, libgen and sci-hub are added (#637)
* Create tmdb

The Movie Database (TMDB) is a popular database for movies and TV shows.

* Create libgen

Library Genesis (Libgen) is a file-sharing based shadow library website for scholarly journal articles, academic and general-interest books, images, comics, audiobooks, and magazines. The site enables free access to content that is otherwise paywalled or not digitized elsewhere.

* Create sci-hub

Sci-Hub is a shadow library website that provides free access to millions of research papers and books by bypassing publishers' paywalls in various ways.

* Update category-entertainment

tmdb was added to this category.

* Update geolocation-!cn

sci-hub and libgen were added.